http unit tests moved to their own class, added unit test for new params parameter, updated http error unit test to cover the new error type and attrs supported now
